Refactor app.py into a modular structure
This commit refactors the main app.py file by organizing the codebase into a more modular and maintainable structure.

Key changes include:
- Created a `config` directory for database configurations.
- Created a `utils` directory for database helpers, date utilities, and transaction mapping functions.
- Created a `blueprints` directory to house API endpoints, with each major functional area (auth, users, medicines, inventory, dispense, goods_received, requisitions, dashboard) separated into its own blueprint file.
- The main `app.py` file has been streamlined to primarily handle Flask app initialization and blueprint registration.
- Imports have been updated across all affected files to reflect the new structure.
- Logging now uses `current_app.logger` within blueprints.

This restructuring improves code organization, making it easier to navigate, develop, and test individual components of your application.
