+++ config/database.py
@@ -1,0 +1,4 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
